requirements.txt
Django==1.10.1
Pillow==2.3.0
django-contrib-comments==1.7.3
django-hitcount==1.2.2
django-model-utils==2.6
django-multiupload==0.5.1
django-notifications-hq==1.2
django-sendgrid==1.0.1
django-url-tools==0.0.8
html5lib==0.999
httplib2==0.8
jsonfield==1.0.3
psycopg2==2.4.5
pyOpenSSL==0.13
pycrypto==2.6.1
python-http-client==2.2.1
requests==2.2.1
sendgrid-django==2.0.0
sorl-thumbnail==12.3
dj-database-url==0.4.1
gunicorn==19.6.0

